+++ python/pathway/tests/temporal/test_operators.py
@@ -1,5 +1,7 @@
 # Copyright © 2024 Pathway
 
+import pytest
+
 import pathway as pw
 from pathway.tests.utils import assert_stream_equality_wo_index
 
@@ -46,7 +48,8 @@ def test_forget_instance():
     assert_stream_equality_wo_index(res, expected)
 
 
-def test_forget_no_instance():
+@pytest.mark.parametrize("public", [True, False])
+def test_forget_no_instance(public: bool):
     t = pw.debug.table_from_markdown(
         """
         a |  t | __time__
@@ -62,11 +65,14 @@ def test_forget_no_instance():
     """
     )
 
-    res = t._forget(
-        pw.this.t + 2,
-        pw.this.t,
-        mark_forgetting_records=False,
-    )
+    if public:
+        res = t.forget(pw.this.t, 2)
+    else:
+        res = t._forget(
+            pw.this.t + 2,
+            pw.this.t,
+            mark_forgetting_records=False,
+        )
     expected = pw.debug.table_from_markdown(
         """
         a |  t | __time__ | __diff__
@@ -83,6 +89,39 @@ def test_forget_no_instance():
         2 |  9 |     8    |     1
         2 |  8 |    10    |    -1
         2 |  9 |    10    |    -1
+    """
+    )
+    assert_stream_equality_wo_index(res, expected)
+
+
+def test_forget_non_append_only():
+    t = pw.debug.table_from_markdown(
+        """
+          | x | __time__ | __diff__
+        1 | 1 |     2    |     1
+        2 | 5 |     4    |     1
+        3 | 5 |     6    |     1
+        1 | 1 |     8    |    -1
+        4 | 2 |    10    |     1
+        4 | 2 |    12    |    -1
+        3 | 5 |    14    |    -1
+        5 | 8 |    16    |     1
+        6 | 8 |    18    |     1
+    """
+    )
+
+    res = t._forget(pw.this.x + 2, pw.this.x, mark_forgetting_records=False)
+    expected = pw.debug.table_from_markdown(
+        """
+          | x | __time__ | __diff__
+        1 | 1 |     2    |     1
+        2 | 5 |     4    |     1
+        3 | 5 |     6    |     1
+        1 | 1 |     6    |    -1
+        3 | 5 |    14    |    -1
+        5 | 8 |    16    |     1
+        6 | 8 |    18    |     1
+        2 | 5 |    18    |    -1
     """
     )
     assert_stream_equality_wo_index(res, expected)
@@ -126,7 +165,8 @@ def test_buffer_instance():
     assert_stream_equality_wo_index(res, expected)
 
 
-def test_buffer_no_instance():
+@pytest.mark.parametrize("public", [True, False])
+def test_buffer_no_instance(public: bool):
     t = pw.debug.table_from_markdown(
         """
         a |  t | __time__
@@ -141,11 +181,13 @@ def test_buffer_no_instance():
         3 |  1 |     6
     """
     )
-
-    res = t._buffer(
-        pw.this.t + 2,
-        pw.this.t,
-    )
+    if public:
+        res = t.buffer(pw.this.t, 2)
+    else:
+        res = t._buffer(
+            pw.this.t + 2,
+            pw.this.t,
+        )
     expected = pw.debug.table_from_markdown(
         """
         a |  t | __time__
@@ -158,6 +200,38 @@ def test_buffer_no_instance():
         2 |  8 |     8
         2 |  9 |     8
         3 |  1 |     6
+    """
+    )
+    assert_stream_equality_wo_index(res, expected)
+
+
+def test_buffer_non_append_only():
+    t = pw.debug.table_from_markdown(
+        """
+          | x | __time__ | __diff__
+        1 | 1 |     2    |     1
+        2 | 5 |     4    |     1
+        3 | 5 |     6    |     1
+        1 | 1 |     8    |    -1
+        4 | 2 |    10    |     1
+        4 | 2 |    12    |    -1
+        3 | 5 |    14    |    -1
+        5 | 8 |    16    |     1
+        6 | 8 |    18    |     1
+    """
+    )
+
+    res = t.buffer(pw.this.x, 2)
+    expected = pw.debug.table_from_markdown(
+        """
+          | x |             __time__ | __diff__
+        1 | 1 |                    4 |     1
+        1 | 1 |                    8 |    -1
+        4 | 2 |                   10 |     1
+        4 | 2 |                   12 |    -1
+        2 | 5 |                   16 |     1
+        5 | 8 | 18446744073709551614 |     1
+        6 | 8 | 18446744073709551614 |     1
     """
     )
     assert_stream_equality_wo_index(res, expected)
@@ -199,7 +273,8 @@ def test_freeze_instance():
     assert_stream_equality_wo_index(res, expected)
 
 
-def test_freeze_no_instance():
+@pytest.mark.parametrize("public", [True, False])
+def test_freeze_no_instance(public: bool):
     t = pw.debug.table_from_markdown(
         """
         a |  t | __time__
@@ -215,10 +290,13 @@ def test_freeze_no_instance():
     """
     )
 
-    res = t._freeze(
-        pw.this.t + 2,
-        pw.this.t,
-    )
+    if public:
+        res = t.ignore_late(pw.this.t, 2)
+    else:
+        res = t._freeze(
+            pw.this.t + 2,
+            pw.this.t,
+        )
     expected = pw.debug.table_from_markdown(
         """
         a |  t | __time__