+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Added
+- `pw.Table.forget` to remove old (in terms of event time) entries from the pipeline.
+- `pw.Table.buffer`, a stateful buffering operator that delays entries until `time_column <= max(time_column) - threshold` condition is met.
+- `pw.Table.ignore_late` to filter out old (in terms of event time) entries.
+
 ## [0.26.0]
 
 ### Added